Abstract algorithm class IncrementalMapperImpl (#3040)
* Abstract struct ``RegistrationStatistics`` to simplify the
IncrementalMapper header
* Abstract an algorithm class as discussed and move in some
indepependent methods

---------
