+++ src/colmap/sfm/CMakeLists.txt
@@ -33,6 +33,7 @@ set(FOLDER_NAME "sfm")
 COLMAP_ADD_LIBRARY(
     NAME colmap_sfm
     SRCS
+        incremental_mapper_impl.h incremental_mapper_impl.cc
         incremental_mapper.h incremental_mapper.cc
         incremental_triangulator.h incremental_triangulator.cc
         observation_manager.h observation_manager.cc
